diesel_attribute_parser::util

Function parse_eq_type

pub fn parse_eq_type(input: ParseStream<'_>, help: &str) -> Result<Type>

Specialized version of parse_eq for syn::Type with a customized error message for readability. This is useful because a great variety of tokens would be valid to parse as a syn::Type.